Refactor SD path and Play Store intent
