+++ app/src/main/java/org/ole/planet/myplanet/datamanager/ConfigurationManager.kt
@@ -21,7 +21,7 @@ import org.ole.planet.myplanet.utilities.LocaleHelper
 import org.ole.planet.myplanet.utilities.NetworkUtils.extractProtocol
 import org.ole.planet.myplanet.utilities.ServerUrlMapper
 import org.ole.planet.myplanet.utilities.UrlUtils
-import org.ole.planet.myplanet.utilities.Utilities
+import org.ole.planet.myplanet.utilities.IntentUtils
 
 class ConfigurationManager(
     private val context: Context,
@@ -225,7 +225,7 @@ class ConfigurationManager(
             builder.setCancelable(true)
             builder.setNegativeButton(R.string.okay) { dialog: DialogInterface, _: Int ->
                 if (playStoreRedirect) {
-                    Utilities.openPlayStore()
+                    IntentUtils.openPlayStore(context)
                 }
                 dialog.cancel()
             }
