+++ app/src/main/java/org/ole/planet/myplanet/ui/sync/SyncActivity.kt
@@ -94,6 +94,7 @@ import org.ole.planet.myplanet.utilities.NotificationUtil.cancelAll
 import org.ole.planet.myplanet.utilities.ServerConfigUtils
 import org.ole.planet.myplanet.utilities.SharedPrefManager
 import org.ole.planet.myplanet.utilities.UrlUtils
+import org.ole.planet.myplanet.utilities.FileUtils
 import org.ole.planet.myplanet.utilities.Utilities
 import org.ole.planet.myplanet.utilities.Utilities.getRelativeTime
 
@@ -234,7 +235,7 @@ abstract class SyncActivity : ProcessUserDataActivity(), SyncListener, CheckVers
     }
 
     private fun clearInternalStorage() {
-        val myDir = File(Utilities.SD_PATH)
+        val myDir = File(FileUtils.SD_PATH)
         if (myDir.isDirectory) {
             val children = myDir.list()
             if (children != null) {
